+++ lib/eve_online.rb
@@ -53,6 +53,7 @@ require 'eve_online/esi/corporation_blueprints'
 
 require 'eve_online/esi/alliances'
 require 'eve_online/esi/alliance'
+require 'eve_online/esi/alliance_corporations'
 require 'eve_online/esi/alliance_icon'
 
 require 'eve_online/esi/dogma_attributes'